Refactor: break down monolithic file into smaller modules for clarity and maintainability
