+++ src/ModUpdatePrompt.py
@@ -1,0 +1,38 @@
+import tkinter as tk
+from tkinter import ttk
+import webbrowser
+
+import Util
+
+class ModUpdatePromptWindow(tk.Toplevel):
+    def __init__(self, parent, mod_page, mod_name, current_version, new_version):
+        super().__init__(parent)
+
+        self.title("Update Available")
+        self.resizable(False, False)
+        
+        ttk.Label(
+            self,
+            text=f"A new version of {mod_name} is available!\n\n"
+                 f"Current version: v{current_version}\n"
+                 f"New version: v{new_version}"
+        ).pack(padx=12, pady=(12, 6))
+        
+        btn_frame = ttk.Frame(self)
+        btn_frame.pack(padx=12, pady=(0, 12))
+        
+        def on_update():
+            webbrowser.open(mod_page)
+            self.destroy()
+            
+        def on_ignore():
+            self.destroy()
+            
+        ttk.Button(btn_frame, text="Update", command=on_update)\
+            .pack(side=tk.LEFT, padx=(0, 6))
+        ttk.Button(btn_frame, text="Ignore", command=on_ignore)\
+            .pack(side=tk.LEFT)
+            
+        self.attributes("-topmost", True)
+        
+        Util.center_window(self)
