+++ src/Settings.py
@@ -1,0 +1,79 @@
+import tkinter as tk
+from tkinter import ttk, filedialog
+import os
+
+import Util
+from Credits import CreditsWindow
+import Config
+
+class SettingsWindow(tk.Toplevel):
+    def __init__(self, parent):
+        super().__init__(parent)
+
+        self.config = Config.config
+
+        self.withdraw()
+        
+        frame = ttk.Frame(self, padding=10)
+        frame.pack(fill=tk.BOTH, expand=True)
+
+        self.show_logs_var = tk.BooleanVar(
+            value=self.config.get("show_cmd_logs", False)
+        )
+        chk = ttk.Checkbutton(
+            frame,
+            text="Show console logs",
+            variable=self.show_logs_var,
+            command=self.on_toggle_logs
+        )
+        chk.grid(row=0, column=0, columnspan=2, sticky="w", pady=(0,8))
+
+        ttk.Label(frame, text="Game Directory:").grid(row=1, column=0, sticky="w")
+        self.game_root_var = tk.StringVar(value=self.config["game_root"])
+        entry = ttk.Entry(frame, textvariable=self.game_root_var, width=50, state="readonly")
+        entry.grid(row=1, column=1, sticky="we", padx=(5,0))
+        pick_btn = ttk.Button(frame, text="...", width=3, command=self.on_change_game_root)
+        pick_btn.grid(row=1, column=2, sticky="e", padx=(5,0))
+        frame.columnconfigure(1, weight=1)
+        ttk.Button(frame, text="Credits", command=lambda: self.open_credits())\
+        .grid(row=2, column=0, columnspan=3, pady=(12,0))
+
+        self.title("Settings")
+        self.geometry("500x120")
+        self.resizable(False, False)
+        self.update_idletasks()
+        Util.center_window(self)
+        self.deiconify()
+        self.transient(parent)
+        self.grab_set()
+
+
+    def open_credits(self):
+        CreditsWindow(self)
+
+    # Something is going wrong here
+    def on_toggle_logs(self):
+        enabled = self.show_logs_var.get()
+        self.config["show_cmd_logs"] = enabled
+        Config.save_config(self.config)
+        if enabled:
+            Config.show_console()
+        else:
+            Config.hide_console() 
+    
+
+    def on_change_game_root(self):
+        new_root = filedialog.askdirectory(
+            title="Select Crossworlds Install Folder",
+            initialdir=self.config["game_root"]
+        )
+        if not new_root:
+            return
+        self.game_root_var.set(new_root)
+        self.config["game_root"] = new_root
+        self.config["game_mods_folder"] = os.path.join(new_root, "UNION", "Content", "Paks", "~mods")
+        Config.save_config(self.config)
+        global GAME_ROOT, GAME_EXE
+        GAME_ROOT  = new_root
+        GAME_EXE   = os.path.join(GAME_ROOT, "SonicRacingCrossWorlds.exe")
+        print("Updated root folder")
